+++ SerenityLib/include/serenity/lua/LuaState.h
@@ -17,7 +17,7 @@ class LUAState {
 
         [[nodiscard]] inline operator lua_State*() { return L; }
 
-        int runFile(const std::string& filename, size_t scriptID) const noexcept;
+        int runFile(const std::string& filename, uint32_t scriptID) const noexcept;
 };
 
 #endif
